Module shuffle::shared

Structs

DevApiClient

Home

Network

NetworksConfig

ProjectConfig

Constants

MAIN_PKG_PATH

Functions

build_move_package

Builds a package using the move package system.

generate_typescript_libraries

Generates the typescript bindings for the main Move package based on the embedded diem types and Move stdlib. Mimics much of the transaction_builder_generator’s CLI except with typescript defaults and embedded content, as opposed to repo directory paths.

get_filtered_envs_for_deno

get_home_path

get_shuffle_project_path

Checks the current directory, and then parent directories for a Shuffle.toml file to indicate the base project directory.

normalized_network

normalized_project_path

read_project_config

send_transaction

Send a transaction to the blockchain through the blocking client.