Module shuffle::shared

Structs

Home

Network

NetworkHome

NetworksConfig

ProjectConfig

Constants

LATEST_USERNAME

LOCALHOST_NAME

MAIN_PKG_PATH

PLACEHOLDER_ADDRESS

Temporary address for initial codegen, replaced on subsequent commands when accounts are available.

TEST_USERNAME

Functions

build_move_package

Builds a package using the move package system.

codegen_typescript_libraries

Generates the typescript bindings for the main Move package. Requires a publishing address for the code generation of script functions that need the address as part of the Module Id. ie: 0x1::MyModule::script_function

get_filtered_envs_for_deno

get_home_path

get_shuffle_project_path

Checks the current directory, and then parent directories for a Shuffle.toml file to indicate the base project directory.

inject_publishing_address_into_manifest

normalized_network_name

normalized_network_url

normalized_project_path

read_project_config